process/perf/params: fix narrowing conversions, type mismatches, and leaks
- process: use std::stoi instead of std::stoull for tgid; avoids
  implicit narrowing from unsigned long long to int
- timer: fix map key type unsigned long -> uint64_t to match
  find_create_timer parameter type
- perf: change trace_type from unsigned int to int so -1 sentinel
  is representable and attr.config <= 0 guard works correctly;
  add perf_fd < 0 guard in stop() before ioctl
- persistent: fix result_bundle leak in load_results; always delete
  the last-allocated unsaved bundle after the parse loop

+++ src/process/process.cpp
@@ -104,7 +104,7 @@ process::process(const std::string &_comm, int _pid, int _tid) : power_consumer(
 					size_t pos = line.find(':');
 					if (pos != std::string::npos) {
 						try {
-							tgid = std::stoull(line.substr(pos + 1));
+							tgid = std::stoi(line.substr(pos + 1));
 						} catch (...) {}
 						break;
 					}
